[Ldap] Add security LdapUser and provider
