+++ src/Symfony/Component/Security/Core/composer.json
@@ -25,13 +25,14 @@
         "symfony/event-dispatcher": "^4.3",
         "symfony/expression-language": "^3.4|^4.0|^5.0",
         "symfony/http-foundation": "^3.4|^4.0|^5.0",
-        "symfony/ldap": "^3.4|^4.0|^5.0",
+        "symfony/ldap": "^4.4|^5.0",
         "symfony/validator": "^3.4|^4.0|^5.0",
         "psr/log": "~1.0"
     },
     "conflict": {
         "symfony/event-dispatcher": "<4.3|>=5",
-        "symfony/security-guard": "<4.3"
+        "symfony/security-guard": "<4.3",
+        "symfony/ldap": "<4.4"
     },
     "suggest": {
         "psr/container-implementation": "To instantiate the Security class",
