+++ src/Symfony/Component/Security/composer.json
@@ -37,11 +37,12 @@
         "symfony/routing": "^3.4|^4.0|^5.0",
         "symfony/validator": "^3.4|^4.0|^5.0",
         "symfony/expression-language": "^3.4|^4.0|^5.0",
-        "symfony/ldap": "^3.4|^4.0|^5.0",
+        "symfony/ldap": "^4.4|^5.0",
         "psr/log": "~1.0"
     },
     "conflict": {
-        "symfony/event-dispatcher": ">=5"
+        "symfony/event-dispatcher": ">=5",
+        "symfony/ldap": "<4.4"
     },
     "suggest": {
         "psr/container-implementation": "To instantiate the Security class",
